Reorganize Polio tests
