+++ plugins/polio/tests/api/campaigns/test_campaigns_1.py
@@ -1,10 +1,9 @@
 import datetime
 
-from django.contrib.auth.models import Permission, User
 from django.utils import timezone
+
 from rest_framework.test import APIClient
 
-from hat.audit.models import Modification
 from iaso import models as m
 from iaso.models import Account
 from iaso.test import APITestCase
@@ -708,125 +707,3 @@ class PolioAPITestCase(APITestCase, PolioTestCaseMixin):
         self.assertEqual(response.status_code, 200, response.content)
         response_data = response.json()
         self.assertEqual(len(response_data), 0)
-
-
-class PreparednessAPITestCase(APITestCase):
-    data_source: m.DataSource
-    source_version_1: m.SourceVersion
-    account: Account
-
-    @classmethod
-    def setUpTestData(cls):
-        cls.data_source = m.DataSource.objects.create(name="Default source")
-        cls.now = timezone.now()
-        cls.source_version_1 = m.SourceVersion.objects.create(data_source=cls.data_source, number=1)
-        cls.account = Account.objects.create(name="polio", default_version=cls.source_version_1)
-        cls.user = cls.create_user_with_profile(username="user", account=cls.account, permissions=["iaso_forms"])
-
-    def setUp(self):
-        """Make sure we have a fresh client at the beginning of each test"""
-        self.client = APIClient()
-        self.client.force_authenticate(self.user)
-
-    def test_two_campaign_round_empty(self):
-        type, created = CampaignType.objects.get_or_create(name=CampaignType.POLIO)
-        campaign_a = Campaign.objects.create(obr_name="campaign A", account=self.account)
-        campaign_a.campaign_types.add(type)
-        campaign_a.rounds.create(number=1)
-        campaign_a.rounds.create(number=3)
-        Campaign.objects.create(obr_name="campaign B", account=self.account)
-        Campaign.objects.create(obr_name="campaign c", account=self.account)
-
-        response = self.client.get(f"/api/polio/campaigns/{campaign_a.id}/", format="json")
-
-        r = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(r["rounds"]), 2)
-
-        response = self.client.get(f"/api/polio/preparedness_dashboard/", format="json")
-        r = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(r), 0)
-
-    def test_two_campaign_round_error(self):
-        type, created = CampaignType.objects.get_or_create(name=CampaignType.POLIO)
-        campaign_a = Campaign.objects.create(obr_name="campaign A", account=self.account)
-        campaign_a.campaign_types.add(type)
-        round_one = campaign_a.rounds.create(number=1)
-        round_three = campaign_a.rounds.create(number=3)
-        campaign_b = Campaign.objects.create(obr_name="campaign B", account=self.account)
-        campaign_c = Campaign.objects.create(obr_name="campaign c", account=self.account)
-        campaign_b.campaign_types.add(type)
-        campaign_c.campaign_types.add(type)
-        response = self.client.get(f"/api/polio/campaigns/{campaign_a.id}/", format="json")
-
-        r = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(r["rounds"]), 2)
-
-        round_one.preparedness_spreadsheet_url = "https://docs.google.com/spreadsheets/d/1"
-        round_one.save()
-        round_three.preparedness_spreadsheet_url = "https://docs.google.com/spreadsheets/d/1"
-        round_three.save()
-
-        response = self.client.get(f"/api/polio/preparedness_dashboard/", format="json")
-        r = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(r), 2)
-        for campaign_round in r:
-            self.assertEqual(campaign_round["status"], "not_sync")
-
-        """[{'campaign_id': '6ecda204-2206-4ae2-a38a-5359684dccf6', 'campaign_obr_name': 'campaign A',
-          'indicators': {}, 'round': 'Round1', 'round_id': 44, 'round_start': None, 'round_end': None,
-          'status': 'not_sync', 'details': 'This spreadsheet has not been synchronised yet'}, {
-                'campaign_id': '6ecda204-2206-4ae2-a38a-5359684dccf6', 'campaign_obr_name': 'campaign A',
-                'indicators': {}, 'round': 'Round3', 'round_id': 45, 'round_start': None, 'round_end': None,
-                'status': 'not_sync', 'details': 'This spreadsheet has not been synchronised yet'}]"""
-
-
-class TeamAPITestCase(APITestCase):
-    fixtures = ["user.yaml"]
-    c: Campaign
-    user: User
-    country: OrgUnit
-    country2: OrgUnit
-
-    @classmethod
-    def setUpTestData(cls) -> None:
-        cls.user = User.objects.get(username="test")
-        cls.country = OrgUnit.objects.create(name="Country A")
-        cls.country2 = OrgUnit.objects.create(name="Country B")
-
-        cls.c = Campaign.objects.create(
-            obr_name="test campaign", account=cls.user.iaso_profile.account, country=cls.country
-        )
-
-    def test_audit_list(self):
-        """Mod a campaign, user can see modification. Limit to another country modification cannot be listed anymore
-        Give access to country in campaign , they can be listed again
-        """
-        self.client.force_authenticate(self.user)
-        self.assertEqual(self.user.is_superuser, False)
-        self.user.user_permissions.add(Permission.objects.get(codename="iaso_polio"))
-        payload = {"obr_name": "test2"}
-        response = self.client.put(f"/api/polio/campaigns/{self.c.id}/", payload, format="json")
-        self.assertJSONResponse(response, 200)
-
-        self.assertEqual(Modification.objects.count(), 1)
-        response = self.client.get(f"/api/logs/?objectId={self.c.id}&contentType=polio.campaign&limit=10")
-        j = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(j["list"]), 1)
-
-        # limit user to the other country. Cannot list
-        p = self.user.iaso_profile
-        p.org_units.set([self.country2])
-        p.save()
-        response = self.client.get(f"/api/logs/?objectId={self.c.id}&contentType=polio.campaign&limit=10")
-        j = self.assertJSONResponse(response, 401)
-        self.assertEqual(j, {"error": "Unauthorized"})
-
-        # limit user to the other country. Cannot list
-        p = self.user.iaso_profile
-        p.org_units.set([self.c.country])
-        p.save()
-        self.client.force_authenticate(self.user)
-
-        response = self.client.get(f"/api/logs/?objectId={self.c.id}&contentType=polio.campaign&limit=10")
-        j = self.assertJSONResponse(response, 200)
-        self.assertEqual(len(j["list"]), 1)
