+++ plugins/polio/tests/api/campaigns/test_campaigns_calculator.py
@@ -1,0 +1,22 @@
+import json
+
+from iaso.test import TestCase
+from plugins.polio.preparedness.calculator import get_preparedness_score
+
+
+class CampaignCalculatorTestCase(TestCase):
+    def setUp(self) -> None:
+        with open("./plugins/polio/preparedness/test_data/example1.json", "r") as json_data:
+            self.preparedness_preview = json.load(json_data)
+
+    def test_national_score(self):
+        result = get_preparedness_score(self.preparedness_preview)
+        self.assertEqual(result["national_score"], 93)
+
+    def test_regional_score(self):
+        result = get_preparedness_score(self.preparedness_preview)
+        self.assertEqual(result["regional_score"], 68.4)
+
+    def test_district_score(self):
+        result = get_preparedness_score(self.preparedness_preview)
+        self.assertAlmostEqual(result["district_score"], 56.25)
