+++ plugins/polio/tests/api/test_lqasim_polio.py
@@ -1,0 +1,139 @@
+import datetime
+from unittest.mock import patch
+
+from django.contrib.gis.geos import MultiPolygon, Point, Polygon
+
+from iaso import models as m
+from iaso.test import APITestCase
+from plugins.polio.api.campaigns.campaigns import CampaignSerializer
+from plugins.polio.preparedness.spreadsheet_manager import *
+
+
+class LQASIMPolioTestCase(APITestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.star_wars = star_wars = m.Account.objects.create(name="Star Wars")
+        marvel = m.Account.objects.create(name="MCU")
+        cls.project = m.Project.objects.create(
+            name="Hydroponic gardens", app_id="stars.empire.agriculture.hydroponics", account=star_wars
+        )
+        sw_source = m.DataSource.objects.create(name="Evil Empire")
+        sw_source.projects.add(cls.project)
+        cls.sw_source = sw_source
+        sw_version_1 = m.SourceVersion.objects.create(data_source=sw_source, number=1)
+        sw_version_2 = m.SourceVersion.objects.create(data_source=sw_source, number=2)
+        star_wars.default_version = sw_version_1
+        star_wars.save()
+
+        cls.jedi_squad = m.OrgUnitType.objects.create(name="Jedi Squad", short_name="Jds")
+
+        cls.jedi_council = m.OrgUnitType.objects.create(name="Jedi Council", short_name="Cnc")
+        cls.jedi_council.sub_unit_types.add(cls.jedi_squad)
+
+        cls.mock_multipolygon = MultiPolygon(Polygon([[-1.3, 2.5], [-1.7, 2.8], [-1.1, 4.1], [-1.3, 2.5]]))
+        cls.mock_point = Point(x=4, y=50, z=100)
+
+        cls.elite_group = m.Group.objects.create(name="Elite councils")
+        cls.unofficial_group = m.Group.objects.create(name="Unofficial Jedi councils")
+        cls.another_group = m.Group.objects.create(name="Another group")
+
+        cls.elite_group = m.Group.objects.create(name="Elite councils", source_version=sw_version_1)
+        cls.unofficial_group = m.Group.objects.create(name="Unofficial Jedi councils")
+        cls.another_group = m.Group.objects.create(name="Another group")
+
+        cls.jedi_council_corruscant = m.OrgUnit.objects.create(
+            org_unit_type=cls.jedi_council,
+            version=sw_version_1,
+            name="Corruscant Jedi Council",
+            geom=cls.mock_multipolygon,
+            simplified_geom=cls.mock_multipolygon,
+            catchment=cls.mock_multipolygon,
+            location=cls.mock_point,
+            validation_status=m.OrgUnit.VALIDATION_VALID,
+            source_ref="PvtAI4RUMkr",
+        )
+        cls.jedi_council_corruscant.groups.set([cls.elite_group])
+
+        cls.jedi_council_endor = m.OrgUnit.objects.create(
+            org_unit_type=cls.jedi_council,
+            version=sw_version_1,
+            name="Endor Jedi Council",
+            geom=cls.mock_multipolygon,
+            simplified_geom=cls.mock_multipolygon,
+            catchment=cls.mock_multipolygon,
+            location=cls.mock_point,
+            validation_status=m.OrgUnit.VALIDATION_VALID,
+        )
+        cls.jedi_squad_endor = m.OrgUnit.objects.create(
+            parent=cls.jedi_council_endor,
+            org_unit_type=cls.jedi_squad,
+            version=sw_version_1,
+            name="Endor Jedi Squad 1",
+            geom=cls.mock_multipolygon,
+            simplified_geom=cls.mock_multipolygon,
+            catchment=cls.mock_multipolygon,
+            location=cls.mock_point,
+            validation_status=m.OrgUnit.VALIDATION_VALID,
+            source_ref="F9w3VW1cQmb",
+        )
+        cls.jedi_squad_endor = m.OrgUnit.objects.create(
+            parent=cls.jedi_council_endor,
+            org_unit_type=cls.jedi_squad,
+            version=sw_version_1,
+            name="Endor Jedi Squad 1",
+            geom=cls.mock_multipolygon,
+            simplified_geom=cls.mock_multipolygon,
+            catchment=cls.mock_multipolygon,
+            location=cls.mock_point,
+            validation_status=m.OrgUnit.VALIDATION_VALID,
+        )
+
+        cls.jedi_council_brussels = m.OrgUnit.objects.create(
+            org_unit_type=cls.jedi_council,
+            version=sw_version_2,
+            name="Brussels Jedi Council",
+            geom=cls.mock_multipolygon,
+            simplified_geom=cls.mock_multipolygon,
+            catchment=cls.mock_multipolygon,
+            location=cls.mock_point,
+            validation_status=m.OrgUnit.VALIDATION_VALID,
+        )
+
+        cls.yoda = cls.create_user_with_profile(username="yoda", account=star_wars, permissions=["iaso_org_units"])
+        cls.luke = cls.create_user_with_profile(
+            username="luke", account=star_wars, permissions=["iaso_org_units"], org_units=[cls.jedi_council_endor]
+        )
+        cls.raccoon = cls.create_user_with_profile(username="raccoon", account=marvel, permissions=["iaso_org_units"])
+
+        cls.form_1 = m.Form.objects.create(name="Hydroponics study", period_type=m.MONTH, single_per_period=True)
+
+        cls.create_form_instance(
+            form=cls.form_1, period="202001", org_unit=cls.jedi_council_corruscant, project=cls.project
+        )
+
+        cls.create_form_instance(
+            form=cls.form_1, period="202001", org_unit=cls.jedi_council_corruscant, project=cls.project
+        )
+
+        cls.create_form_instance(
+            form=cls.form_1, period="202003", org_unit=cls.jedi_council_corruscant, project=cls.project
+        )
+
+    def test_general_status(self):
+        c = Campaign.objects.create(account=self.star_wars)
+        c.rounds.create(number=1, started_at=datetime.date(2021, 1, 1), ended_at=datetime.date(2021, 1, 2))
+        c.rounds.create(number=2, started_at=datetime.date(2021, 3, 1), ended_at=datetime.date(2021, 3, 2))
+        c.rounds.create(number=3, started_at=datetime.date(2021, 4, 1), ended_at=datetime.date(2021, 4, 20))
+
+        with patch("django.utils.timezone.now", lambda: datetime.datetime(2020, 2, 2, 2, 2, 2)):
+            d = CampaignSerializer(instance=c).data
+            self.assertEqual(d["general_status"], "Preparing")
+        with patch("django.utils.timezone.now", lambda: datetime.datetime(2021, 1, 1, 2, 2, 2)):
+            d = CampaignSerializer(instance=c).data
+            self.assertEqual(d["general_status"], "Round 1 started")
+        with patch("django.utils.timezone.now", lambda: datetime.datetime(2021, 1, 3, 10, 2, 2)):
+            d = CampaignSerializer(instance=c).data
+            self.assertEqual(d["general_status"], "Round 1 ended")
+        with patch("django.utils.timezone.now", lambda: datetime.datetime(2021, 4, 20, 10, 2, 2)):
+            d = CampaignSerializer(instance=c).data
+            self.assertEqual(d["general_status"], "Round 3 started")
